Removed PlaylistPage
Moved PlaylistPage locators and methods to SongsPage for easier maintenance
